REF: move copula out of sandbox (for history)
